Remove vertical padding from the ImageTextButton
This makes the alignment of the inventory buttons in the sidebar look
a little better.

They're still not perfectly aligned because for that we'd need to have
the same text size as the images. The image does a seemingly
pixel-perfect scaling somehow so ideally we'd chose the text size
accordingly.