Remove the dual new|initialised state from `World`
Creating a World now initialises it. Always.

This prevents the distinction between a World that has been created on
the game's startup (and would crash the game if it were used) and
world that's been properly initialised and can be used freely. The
world is now always initialised.

Since we were treating the `new`/`initialised` state as a signal
whether a "new game" was started, we've now introduced a new state
that contains the game "not started"/"in progress"/"ended" property.
This is actually better covered than the two "initialised" and "game
ended" booleans anyway since we care about three situations, not four.