Remove the zmq crate
It's probably bitrotted to hell if it's even in the code still. The
crate is optional but it prevents us from building/testing/checking
all features.

E.g.: `cargo clippy --benches --tests --all-features` was failing on
linux when the underlying system dependency was not installed.

And that's a valuable command to run.