Add the Input struct
Right now it's pretty much a transparent wrapper over Command. But the
plan is to enhance the replay capabilities by marking the frame ID and
later on a mouse position as well so we can basically reply the game
at the same speed, timing etc. as it was being played.

Rather than (or in addition to) the fixed replay speed we're doing
right now.