Make the `state.replay` flag work with the replays again
This removes all the timing durations etc. which were causing issues
when doing replays now.

It's important to have the flag set properly so we're able to
distinguish between a regular game and a replay. This will become
important for the mouse position portion of the replay we'll want to
add next.