Don't generate an extra unused seed and World
Starting a new game has printed out two seeds and two replay files.
This was confusing (which one was the real one when choosing the
replay?) as well as wasteful (one of them was generated but never
used).

So this adds a check if this is the first game started and if so, it
won't generate a new world. Rather, it'll use the one already created.