WIP Fix replays mostly
This moves verifications to `Input`.

Adds a fuckton of logging (that we'll probably want to prune maybe).

And switches timer to use frame ticks rather than durations.

That last bit was what was actually causing the desync (the rest is
diagnostics). Because a real-time duration can different amount of
frames even with the same game state and input.

And since the game waits for animations to finish in many places, that
can mess things up.