Fix replay crash when the player dies
At the end of the game, the replay would crash because the original
playthrough logged a lot of additional inputs post-death compared to
the replay.

The replay expected to end as soon as the player died whereas the log
was being recorded during the final fadeout at least.

Now we only record and replay while the player is still alive.