Make the explosion animation time-independent too
The actual areas covered by the explosion are using
`time.percentage_remaining` and `timer.percentage_elapsed` function.
And those two were basing their values on the elapsed time rather than
the frame count.

So if the frame took a different time on the playthrough and replay,
we'd get a desync because the explosion waves would happen in
different frames.

That's now sorted as we're using the frame count here too.

This (in my testing) is the final desync issue that I've discovered so
far. I had a replay that was about five minutes long, using doses and
food to get rid of monsters and nothing happened.

It even worked on the release mode after being recorded in the debug
mode. I think we're finally good.