Make the Simulation Radius constant
This should be completely independent on the window size (and
therefore `state.map_size`) otherwise replays won't work properly
between different window changes.

This is a tradeoff -- it means that there's going to be a window size
cutoff where you could see more on the screen than that's simulated.

But I think this fits my mental model of the game better -- we've made
a lot of changes to make sure the window size is completely incidental
with no impact on the gameplay.

I was honestly surprised this was still here when I found out.