Only show the graphical tile with shattering explosion
This fixes a bug where if the player overdosed on the Cardinal or
Diagonal dose, both the ASCII and graphical corpses would be visible.

This is because when rendering the background tiles, if there was a
fallback from Graphical to Textual triggered, it would use the
cell's *foreground* as the fallback, rather than the background.

This happened during copy/pasting the background rendering code and as
far as I can tell, only showed in this one situation.