Translate gamepad presses into keyboard presses
This makes the gamepad support a bit less intrusive (the
UI/game-handling code doesn't need to know about yet another input
method). It also forces us to add a proper key arrow handling to
everything.

And it seems to work for the main menu and gameplay (though a lot of
the functionality is not supported yet).