Add gamepad left stick movement controls
This lets the player use the gamepad's stick to do the 8-directional
movement. I've tried to implement this with just tapping the stick in
the right direction but never managed to make it precise enough. There
were times where I thought I've pressed it in a diagonal direction but
the character just moved to e.g. left instead.

In Dose Response, a single misstep can be lethal.

So instead, the stick highlights the tile the player would move
to (similar to the mouse hover in the sidebar) and pressing "A" or
"X" (the "action" button on the pad) will actually do the move.

This feels vastly better, allows for a quick play and it's much more
reliable.