Make inventory usable with arrow/gamepad keys
We can now press the "I" key (or the "West" gamepad button) to have
the inventory take focus. In that state, the Up/Down keys will select
the inventory item and the Enter/South key will "use" it (as if that
button was clicked).

This is one of the last blockers of the main game being fully playable
on the controller. But I haven't tested it that way yet.