sidebar: put all cursor key checking behind inventory_focused
Since the `state.keys.matches_code` action is actually destructive (it
pulls keys out of the queue), for things that operate on the sidebar
being focused, those should have the `state.inventory_focused` check
first.

That was mostly the case except for the `Enter` key. So now we've
moved it all there.