Disable inventory focus on mouse click
If the inventory/sidebar has the keyboard focus (because the player
pressed the `I` key), it gets dismissed again if the player clicks the
mouse.

This feels much more intuitive and in line with what I'm used to for
these "modal" panels than just keeping the focus on there even if the
mouse moved the character or something in the meantime.