release.py: Replace boto with the AWS CLI
Another issue with the CI.

Try passing the `--break-system-packages` flag to pip too

We're getting more and more into a dangerous territory, but maybe this
will work?

See if we have the `aws` CLI tooling in the images

If so, we'll just shell out to it and be done.

Try calling the aws cli from the release binary

This will likely fail but let's inspect the environment and see what
comes out.

Actually pull the AWS code into a code path where I can see them

Run an S3 upload destined to fail

Try to upload the release now

We're using (I think!) the correct destination URI now so this should
work.

If it does, we'll clean the script and merge.

Disable showing progress on s3 upload

typo oops

Move the release upload code behind a switch again