main menu: only process keys and focus when active
Right now, the main menu was processing keys and setting focus always
-- even when another window was displayed on top (and therefore the
main menu was not active).

That seems to have worked well until now when we're trying to
introduce arrow key and focus handling to the Settings window.

None of the keys or focus commands were taking place without this
change.