ci: Add a Rust-based package-release script
This is intended to replace the `release.py` file. I'd like to have
all the CI-related scripts etc. written in Rust. For two reasons:

1. Easier to maintain / share code / etc. if everything's written in
the same language (one reason why Cargo's build scripts are in Rust
too).
2. We don't need to depend on os-level dependency management and
differences

I actually ran into the latter point recently with the `aws` Python package
on MacOS having some trouble and basically had to resort to shelling
out into a binary.

We can keep doing here, but we can also just use Cargo library and
basically have the same ecosystem serve the final product as well as
its accompanying scripts.

It's a separate Cargo project inside the `bin` directory. I've tried
to just having a separate `bin` target within the main Dose Response
`Cargo.toml`, but the dependency situation got really complicated.

Basically, I don't want the game to bring in any dependencies that the
package-release script needs and vice versa. And that's really hard if
not impossible to do (`lib.rs` is being built always so those deps
must always be active and can't be turned off by a feature flag).

I don't love it but it works and it's not that much onerous. I guess
something like `cargo-script` could be utilised here, but for now this
works and is one fewer moving part (should cargo-script ever become
part of Cargo proper though...).

This also adds a `make package-release` target that runs all the
check/build/test jobs and executes the script so we can reproduce all
this locally.

Because of that, the CI definition (`ci.yaml`) file is now greatly
simplified and essentially just runs `make package-release`.