ci: Bring back the x86/64 MacOS build
I don't know when but at some point the `macos-latest` image in Github
Actions CI switched from X86/64 to Arm (the M1 etc. chips).

This makes that explicit and also adds the x86 (for as long as it's
available) so we're building for both architectures.

According to the runner-images repo, the `macos-13` image/os
should be the x86 one.

https://github.com/actions/runner-images