package-release: use zip_add_*_from_path
The Windows packages were producing files with literal slashes in
them. The macos ones were fine. Not sure what exactly is going on, but
since the methods actually have methods that accept a Path directly,
we should use that anyway.

Shorter, cleaner, safer code.

And presumably will fix this issue.