Navigate all the menus using gamepad's left stick too
Since using the stick is the primary way of moving the character via
the gamepad, it felt a bit jarring that you couldn't use the same
stick for navigating the menus and had to switch to the D-pad.

So now we can. It's a bit weirder than the D-pad but since both work,
the player can do whatever feels right to them and just mix & match.

I'm pretty sure this is what a lot/most games I've played do and I
definitely missed it here so yep.