package-release: increase the upload timeout
Occasionally, the periodic tests would fail uploading the build saying
the operation timed out.

This adds more context to the message to see a bit better what's going
on.

But primarily, increases the timeout for the HTML request from the
default 30 seconds to 10 minutes. That should hopefully be enough (and
if not, the test is right to fail).