Implement diagonal movement with gamepad's D-pad
This is an extremely naive approach: if you press both e.g. Up and
Left in the same frame, the game will recognise as a North-West
direction rather than two separate movements.

This actually does what I want, but it's not perfectly reliable. If
the button presses fall in separate frames, they will be treated as
separate presses to possibly game-ending consequences.

As such, I'm keeping this in, but it's not ready for prime-time. E.g.
it's not documented in the Help screen.

I've added a couple of suggestions in the code, but neither feels
great from the onset. I may or may not return to this later.