make: add `steam-deck` target
This will build the game in a container of a Debian release old enough
that its glibc version is supported on the Steam Deck.

We use `podman` to actually create the build in the Debian image.

Seems to be working really solidly, actually.

Haven't tested the result on the deck yet, but at least the build
pipeline seems sound.