Don't overwrite the reply file if it already exists
If we record a replay to a given path (via `--replay-file`) and then
mistakenly re-run the same command rather than asking to replay the
file, the file will no longer be overwritten.

If a replay file exists already, the game will try to generate a new
file instead. If that fails, it won't record the replay at all.