Skip replaying frames with the "same" state
This can speed up the replays significantly. Every time the game
update produces a game `State` that's "the same" as the previous
one (it's not perfect identity but we look at the `World`, `Player`
and the window stack) we just skip rendering that frame.

The game update is still run but basically within the same frame so if
we have e.g. 100 frames resulting in the "same" state, we run all 100
game updates within a single frame before processing and rendering the
next one that's actually different.

Obviously, this is only done on replays and even then only when
`replay-full-speed` is set. So the default replay behaviour is to
recreate each frame perfectly -- at the original speed.

This does have some limitations: there are times when seemingly
nothing happens and yet the replays do actually pause. I have a replay
where I stepped away for about ten minutes while playing the game and
the replay is still stopped too. No idea why yet.

But this is a significant improvement. And I feel if we manage to drop the
framerate and go headless, it might get even faster. Which would be
great if/when we turn the replays into a CI system.