loop_state: only warn on vertex/drawcall capacity when changed
Previously, any time the _initial_ vertex or drawcall capacity was not
enough to hold all the objects in the current frame, we would print a
warning.

That could flood the logs.

Now we only warn when the vectors actually grow. So rather than
comparing them to the initial capacity, we compare them to the current
one.

That still shows it but far less often.