Make the replay even faster
Now we skip the state cloning and comparing (where we actually do
process every single frame -- just not render it).

Instead, we set out a few rules that result in the frame being a no-op
and when we hit those, we'll just skip the rest of the processing
right there and then. So no player or monster processing happens.

And I feel the `State` cloning via serialising & deserialising was
probably slowing things too.

With this, the replays feel pretty much as fast as they possibly could
be. And they're replaying all the four recorded ones (including
victory) flawlessly.

There is a bit of a hack we had to do to stop the victory one from
crashing: when a NPC accompanies a player, the bonus will be set at a
subsequent frame for whatever reason. And the frame where the Victory
"bonus" is set looks otherwise as a frame we should skip.

This is because the whole victory condition was kinda hacked on top of
the existing bonus system.

I don't want to change that because it would invalidate the replay and
honestly I just don't care now.

But it means we're piling another hack on top of this bonus-laden
hack.

Whatever. It works and I honestly don't expect that many new changes
to the game.