Add a "headless" mode
When passing `--headless` to a game replay, the replay will not open a
window and instead just completely "GUI-less". No OpenGL contexts,
Audio etc. will be initialised.

This should be ideal for CI where even if we could open a window and
show graphics, we wouldn't want to.

Plus, this is super fast because there's no forced time delay between
each frame.