Add the `blake3` dependency
This is a fast (faster than all them md5/sha/etc.) hash. It's
cryptographically secure too which we don't need and it means that
it'll take longer than it potentially could, but this is a solid and
speedy hash and why are you even arguing with me here?

Anyways, the plan is to be able to record a hashed `Verification`
instead of the entire deal (or maybe do that out-of-band) to reduce
the size of the replay file.