Add hash-based verification of the game state
Right now, we've switched to just entering the hash of `Verification`
to the replay log and then expecting to get same out.

But it works \o/.

Next we need to clean up the code some (right now we're converting the
Verification into JSON and hashing *that* -- it's slow, ugly, possibly
unreliable (if the json fields get reordered and whatnot)).