Hash `Verification` faster and more reliably
Previously, we've serialised the `Verification` struct to a JSON and
then hashed that as bytes.

This is slow (IDK how much but it is) and more importantly:
unreliable. I don't know whether there are any `json_serde` guarantees
but at least in theory, it's absolutely possible to output equivalent
JSON files that nonetheless aren't byte-wise compatible (to say
nothing of whitespace etc.).

Unfortunately, `blake3::Hasher` doesn't have all the convenience
methods that `std::hash::Hasher` has nor the ability to utilise the
`std::hash::Hash` trait.

So we're converting everything to byte arrays and passing those to
blake3 #thisisfine.