Fix replays with the Double AP bonus
The skip frames weren't working properly because we were only
considering the situation where the `Player` had a single action
point. But they can of course have two.

Also, we were moving through the animations which in a particular
instance messed up the validations.

Now it all works well in all the cases we've discovered so far.