Make package-release into a "standalone" cargo-script file
I never really liked idea of having nested crates in the repo and this
was just a workaround for not having the "cargo script" functionality
fully out.

But it is available on nightly and unless this is a massive pain to
set up in the CI, I'm fine with just using it instead. And at some
point, it should become stable.