settings: keep the keyboard focus on
When navigating to an Settings option using the arrows/dpad etc. and
pressing Enter to flip that option, the keyboard focus used to be
cleared.

Now it's kept. But we still clear any keyboard focus if we flip an
option by mouse-clicking on it or pressing the hotkey.

Ideally, we'd move the focus on that new entry instead, but that can
be done later.