help: fix keyboard scrolling
The previous commit made the keyboard scroll not work.

With this, both work, but it is a bit messy: right now, there's no way
to get the current scroll bar position from egui (that I could find --
on the version we're using).

Which means that when we do start scrolling with the keyboard, we
should be able to update the current scroll position. But instead, we
start from top.

This is hopefully a pretty niche combination (you need to scroll using
the mouse and *then* press the up/down arrows), but it is a bug.

The only workaround I could think of is to convert the arrow up/down
presses to egui's `RawInput` but we can only pass that at the
beginning of a new frame. So that would result in the key press
scrolling a frame later. Not the end of the world necessarily and
possibly worth doing but yeah.