help: align the scrollbars correctly
Before this, their horizontal position depended on the width of the
longest line in each Help window. That means if the lines didn't go
all the way to the right, the scrollbar would sit in, like, the middle
of the window.

With this, they're all in the same position -- completely to the right
edge of the window.