help: implement all scrolling through egui's `RawInput`
Both the mouse wheel scroll as well as the keyboard's `Up` and `Down`
arrows pressed in the Help screen use the same underlying mechanisms
to drive scrolling.

The keyboard events are converted into `scroll_delta` within
`RawInput` and that is what egui uses.

That greatly simplifies the state management -- we only ever store the
delta and just set it or not in a given frame.

It works great and also (as this is handled within egui's scrolling
code) the screens now never go past the top/bottom of the scrolling
position either.