Update Glutin to the latest version (0.32)
It's now using raw-window-handle, there's a glutin-winit crate and
the way Winit works has changed significantly (it calls the loop for
you rather than you iterating over it).

This version now builds, but there are kinks to overcome and it
amounts to a complete rewrite of the `glutin.rs` file.