Use a proper fixed timestep
From now on, regardles of the display refresh rate, the game state is
updated 60 times a minute and rendered in the next available time
window after the update.

This looks and feels perfoct on the M-series MacBook Pro and I'm
assuming it'll work on a regular 60FPS as well.

The slight downside is that due to Winit's owning of the event loop,
we have to run the "fixed time game loop" in a separate thread, but it
seems to work great and it's a very lightweight thread.