Enable all backends in dev but make SDL3 the default in Settings
We want all backends to be enabled in the `dev` profile so we can
catch any errors with them going stale and whatnot.

But we do want to default to the SDL3 one -- at least while we develop
and test it. Which can be controlled by the value set in
`settings.toml`.