Clean up SDL3 loop
Remove a lot of comments and commented-out code. Pull the game
updating/rendering code to a separate function. Remove some warnings.

FIX the game update dt!!

Turns out, we were doing integer conversion wrongly and the
`target_dt` was always set to precisely `16ms`! That's why the always
felt slightly off. Now that we're setting it correctly to
`16.666...` (with whatever precision is available), the numbers make
fare more sense.

And I think this is pretty much it for the loop:

We handle if the game update runs quickly (by having the explicit loop
with small-increment waits to get to the next targeted time to
execute).

We don't wait if vsync or whatever handle timing for us -- we just hop
on to the next update, no artificial waiting required.

And if a frame takes longer, we'll again skip waiting and go directly
to the next one. There shouldn't be any spiralling. We'll just drop
FPS.