Add SDL2 backend
This is just a test to see how/whether this works better than SDL3.

The idea is this: SDL3 has large chunks of the supporting libraries
not implemented. That's not necessarily a problem.

What is (or could be) is that it seems to be all in on the new
graphics stack (vulcan, metal, the whatever DX does). And I'm not
ready to investigate that yet.

Plus SDL2 still works, is still in use. The same argument that's
making me break with winit+glutin is also what should convince me to
stick with SDL2 now instead of SDL3.