Fix SDL2 window opening
We were missing a call to egui_context.begin_pass. Now that works,
SDL2 is actually working. In that you see graphics being rendered, the
game responding to input etc.

The timing is super off on the MacOS, but that's likely because we're
not using the fixed time loop (which set us off on this whole
journey in the first place).