SDL3: Set up OpenGL, render the game
The code is nearly identical to SDL2 (bless), only a couple
differences.

Unfortunately, I'm getting the same freeze here on my linux machine as
with Winit+Glutin. And this has no multithreading or anything like
that.

So, it seems the only real option is good old SDL2 (assuming that
doesn't have the same problem -- if it does, I'll need to investigate
further but I haven't seen it yet).

Anyway, the SDL3 backend is complete as far as I can tell, but it must
be disabled because we can't run both of them at the same time --
there's a runtime crash when both are loaded.