CI: enable the build-script verbosity only on the main build commands
It's super noisy because the `-vv` switch for Cargo has it enabled for
every dependency. That's too much.

So here we build it, then update build.rs to re-run again and then
build it verbose again. That way everything's been built and compiled
and all we should see is the build script output.