Upgrade to Rodio 0.21.1
This upgrade has a few breaking changes, so I took the opportunity to
dig deeper and see if we can simplify our code.

I believe this was a success! There's less exposure to rodio's native
types in the game code, the audio subsystem is easier to set up,
handling `None` I think was reduced too and we're no longer
maintaining and mixing a queue for the sound effects.

The only thing I'm a bit worried about is that we're now opening (and,
to be fair, dropping) a new `Sink` object for every new sound. I seem
to recall that being potentially a problem which led me to doing some
pre-processed mixing before. But maybe that's no longer the case?

I'll need to test that.