audio: rename `mix_sound_effect` to `play_sound`
We're not directly mixing anything now and this indeed just plays the
sound now.

If, in the future, we do decide that this will just either push a
label on a queue or mix some sounds in, the game code intent is still
to simply record that we want to play a sound here, so this name would
stay regardless I reckon.